ref intermediate step: move secondary provider def
