+++ include/libaktualizr/secondary/secondary_provider.h
@@ -5,8 +5,8 @@
 
 #include "libaktualizr/config.h"
 #include "package_manager/packagemanagerinterface.h"
-#include "storage/invstorage.h"
-#include "uptane/tuf.h"
+
+class INvStorage;
 
 class SecondaryProviderBuilder;
