+++ src/libaktualizr/primary/CMakeLists.txt
@@ -9,8 +9,6 @@ set(HEADERS secondary_config.h
             aktualizr_helpers.h
             initializer.h
             reportqueue.h
-            secondary_provider.h
-            secondary_provider_builder.h
             sotauptaneclient.h)
 
 add_library(primary OBJECT ${SOURCES})
