+++ src/virtual_secondary/managedsecondary.h
@@ -8,8 +8,8 @@
 #include <boost/filesystem.hpp>
 #include "json/json.h"
 
-#include "libaktualizr/types.h"
 #include "libaktualizr/secondary/secondaryinterface.h"
+#include "libaktualizr/types.h"
 #include "primary/secondary_config.h"
 
 namespace Primary {
