+++ src/virtual_secondary/partialverificationsecondary.cc
@@ -1,7 +1,7 @@
-#include "libaktualizr/types.h"
-#include "libaktualizr/secondary/secondaryinterface.h"
-#include "logging/logging.h"
 #include "partialverificationsecondary.h"
+#include "libaktualizr/secondary/secondaryinterface.h"
+#include "libaktualizr/types.h"
+#include "logging/logging.h"
 #include "utilities/exceptions.h"
 
 namespace Uptane {
